+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Display character's name in header #3
+
+## [2.4.0] - 2019-03-24
 ### Added
 - Support for multiple characters
 - Support for exporting characters (as YML) and share it with others